requirements.txt
Flask
Flask-Cors
mysql-connector
plotly
python-dotenv